+++ src/databricks/labs/ucx/source_code/linters/pyspark.py
@@ -13,7 +13,7 @@ from databricks.labs.ucx.source_code.base import (
 )
 from databricks.labs.ucx.source_code.linters.imports import ASTLinter, TreeWalker
 from databricks.labs.ucx.source_code.queries import FromTable
-from databricks.labs.ucx.source_code.linters.ast_helpers import AstHelper
+from databricks.labs.ucx.source_code.linters.python_ast import AstHelper
 
 
 @dataclass
